[SD-226] Minimize usage of MonadIO
Problem: MonadIO constraint is used in definition of most AVL+ API
functions, although it is really required only for few of them. This
limits possible usage of AVL+ making it impossible to use in STM monad.

Cause: MonadIO is a part of Base constraint.

Fix: just remove MonadIO from Base and use it in few functions which
really require it.